Bluetooth: Audio: PAC discover by UUID added
Instead of using the read by UUID, we do a "proper"
discover by UUID. The reason for this is that read by
UUID is not mandatory in the BAP specs, and that we need
to do a long read on the PAC records anyhow, and that
cannot be done with the read by type/UUID.

+++ include/zephyr/bluetooth/audio/bap.h
@@ -1236,8 +1236,13 @@ struct bt_bap_unicast_client_discover_params {
 	/** Number of endpoints found */
 	uint8_t num_eps;
 
-	/** Error code. */
-	uint8_t err;
+	/** @brief Error code.
+	 *
+	 * 0 indicates no error.
+	 * A positive value indicates a BT_ATT_ERR_* error.
+	 * A negative value indicates an errno value.
+	 */
+	int err;
 
 	/** Read parameters used interally for discovery */
 	struct bt_gatt_read_params read;
